Удаляет моки отзывов и создает файл utils
