--- project/src/components/card/card.prop.jsx
@@ -1,3 +1,0 @@
-import PropTypes from 'prop-types';
-
-export default PropTypes.number.isRequired;
