--- project/src/mocks/reviews.js
@@ -1,52 +1,0 @@
-const comments = [
-  {
-    comment: 'bad',
-    date: '2019-05-08T14:13:56.569Z',
-    id: 1,
-    rating: 4,
-    user: {
-      avatarUrl: 'img/avatar-angelina.jpg',
-      id: 4,
-      isPro: true,
-      name: 'Max',
-    },
-  },
-  {
-    comment: 'good',
-    date: '2019-05-08T14:13:56.569Z',
-    id: 2,
-    rating: 2,
-    user: {
-      avatarUrl: 'img/avatar-angelina.jpg',
-      id: 4,
-      isPro: false,
-      name: 'Olya',
-    },
-  },
-  {
-    comment: 'exellent',
-    date: '2019-05-08T14:13:56.569Z',
-    id: 3,
-    rating: 5,
-    user: {
-      avatarUrl: 'img/avatar-max.jpg',
-      id: 4,
-      isPro: true,
-      name: 'Vasya',
-    },
-  },
-  {
-    comment: 'nice',
-    date: '2019-05-08T14:13:56.569Z',
-    id: 4,
-    rating: 4,
-    user: {
-      avatarUrl: 'img/avatar-max.jpg',
-      id: 4,
-      isPro: false,
-      name: 'Petya',
-    },
-  },
-];
-
-export default comments;
